+++ plotnine/tests/test_geom_polygon.py
@@ -29,3 +29,12 @@ def test_aesthetics():
                       color='yellow', show_legend=False))
 
     assert p + _theme == 'aesthetics'
+
+
+def test_no_fill():
+    p = (ggplot(df, aes('x', group='factor(z)'))
+         + geom_polygon(aes(y='y'), fill=None, color='red', size=2)
+         + geom_polygon(aes(y='y+2'), fill='None', color='green', size=2)
+         + geom_polygon(aes(y='y+4'), fill='none', color='blue', size=2)
+         )
+    assert p + _theme == 'no_fill'