+++ doc/changelog.rst
@@ -14,6 +14,8 @@ v0.2.1
 
 - Fixed bug in :class:`~plotnine.stat_summary` where the you got
   an exception for some types of the `x` aesthetic values.
+  
+- Fixed bug where ``ggplot(data=df)`` resulted in an exception.
 
 
 v0.2.0